requirements.txt
asgiref==3.5.2
Django==3.2.7
django-cleanup==5.2.0
django-phonenumber-field==5.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
phonenumbers==8.12.31
Pillow==8.3.1
psycopg2==2.9.3
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
